remove k8sutil fork package, move to clientutil
