+++ pkg/operator/clientutil/merge.go
@@ -1,18 +1,4 @@
-// Copyright 2016 The prometheus-operator Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package k8sutil
+package clientutil
 
 import (
 	"encoding/json"
@@ -23,7 +9,10 @@ import (
 	"k8s.io/apimachinery/pkg/util/strategicpatch"
 )
 
-// MergePatchContainers adds patches to base using a strategic merge patch and iterating by container name, failing on the first error
+// MergePatchContainers adds patches to base using a strategic merge patch and
+// iterating by container name, failing on the first error.
+//
+// Copied from github.com/prometheus-operator/prometheus-operator/pkg/k8sutil.
 func MergePatchContainers(base, patches []v1.Container) ([]v1.Container, error) {
 	var out []v1.Container
