--- pkg/operator/k8sutil/README.md
@@ -1,7 +1,0 @@
-# k8sutil
-
-This folder is a temporary fork of the Prometheus Operator k8sutil, needed to
-lower the version dependency on the Kubernetes client. Once all of the Grafana
-Agent's dependencies have been updated to ones that use at least client version
-0.21.0, this package can be removed and replaced with the official package.
-
