--- pkg/operator/k8sutil/k8sutil.go
@@ -1,35 +1,0 @@
-// Copyright 2016 The prometheus-operator Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package k8sutil
-
-import (
-	"regexp"
-	"strings"
-
-	"k8s.io/apimachinery/pkg/util/validation"
-)
-
-var invalidDNS1123Characters = regexp.MustCompile("[^-a-z0-9]+")
-
-// SanitizeVolumeName ensures that the given volume name is a valid DNS-1123 label
-// accepted by Kubernetes.
-func SanitizeVolumeName(name string) string {
-	name = strings.ToLower(name)
-	name = invalidDNS1123Characters.ReplaceAllString(name, "-")
-	if len(name) > validation.DNS1123LabelMaxLength {
-		name = name[0:validation.DNS1123LabelMaxLength]
-	}
-	return strings.Trim(name, "-")
-}
